UC#1 Create filter by population function

+++ src/main/java/com/softserve/peopleai/countries/Country.java
@@ -58,7 +58,7 @@ public class Country {
 
   private Maps maps;
 
-  private int population;
+  private Long population;
 
   private Gini gini;
